Improve syntax highlighting for code blocks
We do not use P4 as the default language for every code block. We now
have extra syntax files for .proto, .cpp and .pb.txt. The C++ syntax
highlighter was taken from the Madoko github. There is probably a lot of
room for improvement for the proto ones.

I added an FAQ to the docs' README to keep track of some Madoko
"specificities".