Clean-up tables in specification
I have given-up on fixing the tables in the generated PDF. It may be
possible by injecting some Tex directly in the Madoko source file, but
it sounds tedious and hard to maintain. So instead, I have cleaned-up
the tables that could be cleaned-up easily (i.e. these that don't have
long lines of text) using the Madoko online editor, and I have replaced
all other tables with lists. I did prefer the tables (in HTML) but the
lists do look okay IMO and they should be easier to modify & maintain.

Fixes #12