Use Madoko bibliography support for references (#28)
* Use Madoko bibliography support for references

Madoko uses BibTeX for bibliography which supports online / web
references. We move all the references to a .bib file and include it in
the Madoko source. The end-result looks very nice and references will
definitely be easier to maintain this way.

It seems that dvipng is used to render "math" inside BibTex strings (we
need it to render P4_16 in subscript notation). For people building
locally, dvipng will need to be installed.

Fixes #25

* Remove tooltip box when hovering on Bib references

Titles were not computed properly by Madoko, si I chose to remove them
altogether using jQuery.