Tentative support for new (user-defined) types in P4Info
We update P4DataTypeSpec to be able to represent user-defined types in
P4Info. Translation (when the control-plane view differes from the
underlying data-plane type) may be required for new types through the
@p4runtime_translation annotation and this information is preserved in
P4Info. As decided previously, we do not break backward-compatibility
for action parameters and match fields. For these we do not use
P4DataTypeSpec but instead we add a new Protobuf field ("type_name")
which is set if the action parameter / match field is a user-defined
type (which must map to an underlying bit vector in this case).