Remove meter color-awareness information from P4Info
There doesn't seem to be any point in keeping this in P4Info at the
moment. First the semantics are murky because in PSA each execute call
on a meter is either color-aware or color-blind (i.e. it is not an
attribute of the meter object itself). Second, at the moment this
information cannot be leveraged in any way by a P4 client since
P4Runtime does not expose any knob to enable / disable color-awareness
on a per meter entry basis at runtime. Even if there was such a knob, it
would be hard to reconciliate it with PSA semantics: in PSA the same
meter can be "executed" in 2 different actions, in color-aware mode in
one and in color-blind mode in the other; these 2 action can use the
same indirect meter entry.

Fixes #58