Better support for @defaultonly and @tableonly annotations
These annotations are now "first-class citizens" in P4Info: an enum
field is used to indicate the scope of the action, i.e. how the action
can be used.

This commit also removes const_default_action_has_mutable_params, which
was introduced for P4_14 initially and has fallen out of use.

Fixes #83