Various small proposed improvements
Including:

Mention that when a write batch completes, operations that return
error status are not committed to the data plane.

Allow `ROLLBACK_ON_ERROR` batch write requests to be implemented by
servers for some batch operations, but not others, similar to how this
is specified for `DATAPLANE_ATOMIC` write requests.