Update the "key" that uniquely identifies a client
I originally wrote only election_id, but I think it is really the
triple of values device_id, role_id, and election_id that is required
to uniquely identify a client.  At least, that is so if election_id
values can be reused for different (device_id, role_id) pairs.