Add type_name field to P4Info Metadata message
Add a description of this field to the text of the section describing
the contents of ControllerPacketMetadata messages.

Also add a description of the type_name field in the ValueSet section.
It was already re-using the MatchField messages used by table search
key fields, and already had this type_name field in it, but it seems
best to explicitly mention that it can be included in MatchField
messages for ValueSet objects, too.