Support wildcard reads for multicast groups and clone sessions
For consistency with the rest of the P4Runtime API, a wildcard Read is
achieved by "setting" the `multicast_group_id` (resp. `session_id`)
field to its default value, which is 0 for sclar fields.

This means that we introduce a small difference with the PSA spec, which
explicitly allows 0 for clone session ids.