Incorporate feedback. Replace maps with repeated ParsedAnnotation.
This more closely preserves the order and content of annotations. Note the effect of shadowing (later kv-pairs replacing earlier ones with the same key) and accumulation mean that you cannot truly reproduce  the source code by reassembling the list of ParsedAnnotations.  If we want to more-closely preserve annotations, we should *not* deprecate the original annotations, in which case this commit needs a follow-on to remove the deprecation extensions.