Add generated Go code for P4Runtime
The generated files are placed under the go/ directory, and the 2 Go
packages which can be imported by other projects are
"github.com/p4lang/p4runtime/go/p4/v1" and
"github.com/p4lang/p4runtime/go/p4/config/v1".

As part of CI, we now check that the generated Go files are up-to-date
with respect to the checked-in Protobuf files. We also check that the
go.mod file is "tidy".

When updating the Protobuf files, a contributor can re-generated the Go
files with `./build/update_go.sh`.

Fixes #260