Add generated rust code
The rust code is generated using the [`neoeinstein/protoc-gen-prost`]
(https://github.com/neoeinstein/protoc-gen-prost),
which leverages the `prost` crate for protobuf and `tonic` crate for
client/server.

The rust crates used are:

pbjson and pbjson-types: 0.7.0
prost: 0.13.1
tonic: 0.12.0

Currently, the generated code does not use `protoc-prost-serde` to
support serde, since serde and json deserialization is not necessary for
the basic use case.

Signed-off-by: Campbell He <kp.campbell.he@duskmoon314.com>