Pin the dependencies of the Rust code gen tools
In the previous commit, only the version of `protoc-gen-prost/tonic` was
pinned. However, Cargo would still ignore the lock file and use the
latest version of dependencies for `protoc-gen-prost/tonic`. This meant
that any new release of dependencies (such as `prost-build`) could
result in differences in the generated code.

This commit addresses the issue by adding `--locked` to the `cargo
install` command, ensuring Cargo respects the lock file when installing
code generation tools.

The Rust code has been regenerated using this updated configuration.

Signed-off-by: Campbell He <kp.campbell.he@duskmoon314.com>